Complete with exercises

+++ albums_migration.sql
@@ -5,7 +5,7 @@ CREATE table albums (
     artist VARCHAR(50) NOT NULL,
     name VARCHAR(100) DEFAULT 'Untitled',
     release_date INT(4),
-    sales FLOAT,
+    sales DECIMAL(6,3),
     genre VARCHAR(20) DEFAULT 'Unknown',
     PRIMARY KEY (id)
 );
